refactored state management 2

+++ src/app/components/EditModal.tsx
@@ -4,19 +4,15 @@ import { Screenshot } from "../types/types";
 import ExistingTags from "./collectionModal/ExisitngTags";
 import { deleteScreenshot } from "../mutations";
 import AddNewTag from "./collectionModal/AddNewTag";
+import { useAuthContext } from "../context/AuthContext";
 
 interface EditModalProps {
   screenshot: Screenshot;
   handleModal: () => void;
   toggleModal: boolean;
-  userId: string;
 }
-function EditModal({
-  screenshot,
-  handleModal,
-  toggleModal,
-  userId,
-}: EditModalProps) {
+function EditModal({ screenshot, handleModal, toggleModal }: EditModalProps) {
+  const { userId } = useAuthContext();
   const { id, siteName, tags } = screenshot;
   return (
     <Modal
@@ -36,8 +32,8 @@ function EditModal({
           </button>
         </div>
 
-        <ExistingTags tags={tags} screenShotId={id} userId={userId} />
-        <AddNewTag screenShotId={id} existingTags={tags} userId={userId} />
+        <ExistingTags tags={tags} screenShotId={id} />
+        <AddNewTag screenShotId={id} existingTags={tags} />
 
         <form action="" className="text-center mt-6">
           <button
@@ -45,7 +41,11 @@ function EditModal({
             className="bg-red-600 px-4 py-2 rounded-md"
             onClick={(e) => {
               e.preventDefault();
-              deleteScreenshot(id, userId);
+              if (userId) {
+                deleteScreenshot(id, userId);
+              } else {
+                console.error("User ID is null. Cannot delete screenshot.");
+              }
             }}
           >
             Delete Screenshot
